chore: document for why setup.py is required
while researching right way for packaging and distributing as per #14, i am currently searching for if calling setup.py to build the wheels n stuff is still preferred way etc. so obviously I wanted to know for what setup.py is actually still needed. and because python packaging is wild its good to document the *true* purpose of each file