refactor: turn Docker's secureappuser into an ARG. ensure COPY'd files are also owned by him
one may say the additionaly added USER and WORKDIR are redudant because they're not being changed in any way anywhere
this is true but i came to like explicit almost-self-contained blocks of stuff and commented regions better
i.e. under "Configure Container Startup" we have WORKDIR and USER
for project-specifc RUN instructions we always have USER in front of it
so if one drops a block with USER root \n RUN he can do that without needing to add a line. each run in project is specific.
why no WORKDIR also for every RUN? because there is an explicit workdir in the region. another region would potentially get its own WORKDIR following this scheme, depending on what it is about